==> src/2_outlier_detection.ipynb <==
# Load imports and data

import numpy as np 
import pandas as pd

data = pd.read_csv('output/loan_dev.csv')
# data = pd.read_csv('output/loan_dev_transformed.csv')
data

# Outlier Detection

- Standard Deviation
- Z-Score
- Interquartile Range (IQR)
- Percentile

for col in data.columns:
    print("-"*50)
    print(f"Statistics of {col}")
    print("-"*50)
    #data[col].plot.box();

    q3, q1 = np.percentile(data[col], [75 ,25])

    print("Min: ", end="");                 print(data[col].min())
    print("Max: ", end="");                 print(data[col].max())
    print("Mean: ", end="");                print(data[col].mean())
    print("Median: ", end="");              print(data[col].median())
    print("Standard deviation: ", end="");  print(data[col].std())
    print("Skewness: ", end="");            print(data[col].skew())  
    print("Variance: ", end="");            print(data[col].var())
    print("Kurtosis: ", end="");            print(data[col].kurtosis())
    print("Interquartile Range: ", end=""); print(q3 - q1, end=" | "); print("Q1: {} | Q3: {}".format(q1,q3))

# var() method returns the unbiased variance
# Medium-High std which indicates that the data is spread out and not clustered around the mean
# Skewness > 1 => it is skewed to the right



## Standard Deviation

outliers_std = {}
for col in data.columns:
    data_mean, data_std = data[col].mean(), data[col].std()
    # identify outliers
    cut_off = data_std * 4
    lower, upper = data_mean - cut_off, data_mean + cut_off 
    outliers_std[col] = pd.concat([data.loc[data[col] < lower] , (data.loc[data[col] > upper])])

for key in outliers_std.keys():
    if (len(outliers_std[key]) > 0):
        print("-"*50)
        print(f"Outliers of {key} - mean: {data[key].mean()} - std: {data[key].std()}")
        print("-"*50)
        print(outliers_std[key][key])

outliers_ids_std = set([val for key in outliers_std.keys() for val in outliers_std[key]['Id']])
outlier_ratio = len(outliers_ids_std) / len(data)
outlier_ratio

## Z-Score

outliers_z = {}
for col in data.columns:
    z_val = (data[col] - data[col].mean()) / data[col].std()
    outliers_z[col] = data.loc[(z_val < -4) | (z_val > 4)]

for key in outliers_z.keys():
    if (len(outliers_z[key]) > 0):
        print("-"*50)
        print(f"Outliers of {key} - mean: {data[key].mean()} - std: {data[key].std()}")
        print("-"*50)
        print(outliers_z[key][key])

outliers_ids_z = set([val for key in outliers_z.keys() for val in outliers_z[key]['Id']])
outlier_ratio = len(outliers_ids_z) / len(data)
outlier_ratio

# Dataset without Outliers

outliers_ids = outliers_ids_std.union(outliers_ids_z)
manual_outlier_df = data.loc[~data['Id'].isin(outliers_ids)]
manual_outlier_df

# Save data

manual_outlier_df.to_csv('output/loan_dev_without_outliers.csv', index=False)
# manual_outlier_df.to_csv('output/loan_dev_transformed_without_outliers.csv', index=False)
